remove extra options, refactor collection
