+++ metricbeat/module/system/service/service_test.go
@@ -24,9 +24,10 @@ import (
 	"time"
 
 	"github.com/coreos/go-systemd/v22/dbus"
-	"github.com/stretchr/testify/assert"
-
 	"github.com/elastic/beats/v7/libbeat/common"
+
+	//mbtest "github.com/elastic/beats/v7/metricbeat/mb/testing"
+	"github.com/stretchr/testify/assert"
 )
 
 var exampleUnits = []dbus.UnitStatus{
